Compiler.Ast

Description

This module contains the AST

Documentation

newtype Identifier

Identifier is a simple identifier, e.g. for a parameter

Constructors

Identifier String

data Name

Name represents a Name of something with the relative path to it

Constructors

Name

Fields

path :: [Identifier]

getIdentifier :: Identifier

data Type

Type are different types of types

Constructors

PrimType PType
Represents a primitive java type

RefType Name
Represents a reference type

JVoid
Represents the void type in java

data PType

PType represent the primitive types

Constructors

Boolean

Int

Char

data Mod

Mod are the different Modifiers in Java

Constructors

Public

Protected

Private

Static

Abstract

data Class

Class is a java class definition

Constructors

Class Identifier [Mod] [Decl]

data Decl

Decl are the different types of declaration inside a class

Constructors

Field VarDecl

Constructor

Method

Fields

getIdentifier :: Identifier

getMods :: [Mod]

getReturnType :: Type

getParamList :: [(Type, Identifier)]

getBody :: Maybe Statement

data VarDecl

VarDecl is used for field definitions and local variable declarations

Constructors

VarDecl

Fields

getIdentifier :: Identifier

getMods :: [Mod]

getType :: Type

getRHS :: Maybe Expression

data Expression

Expression is something that can be evaluated to a value

Constructors

TernaryIf
Short notation if, e.g. someBool? 42 : 1337

Fields

getCond :: Expression

getElseStmt :: Expression

getThenStmt :: Expression

PrimBinOp BinOp Expression Expression
Primitive binary Operation

PrimUnOp UnOp Expression
Primitive unary Operation

This
this keyword

Iden Name
A variable

Select Expression Identifier

Literal Lit
All kind of literals

ExprExprStmt StmtExpr
A StatementExpression that is in an Expression position

Cast Type Expression

data Statement

Statement is dual to Expression since it does not evaluate to a value

Constructors

While

Fields

getCond :: Expression

getBody :: Maybe Statement

If

Fields

getCond :: Expression

getThenStmt :: Maybe Statement

getElseStmt :: Maybe Statement

Block [Statement]

Return (Maybe Expression)

LocalVar VarDecl

StmtExprStmt StmtExpr

data StmtExpr

StmtExpr can be a Statement as well as an Expression

Constructors

Assign AssignOp Name Expression

Instantiation Name [Expression]
Using new

Apply Expression [Expression]

SEUnOp IncrOrDecr Expression
UnOp that returns something and has a side effect

data Lit

Lit are the Java literals

Constructors

IntegerL Integer

BooleanL Bool

CharL Char

StringL String

Null

data AssignOp

AssignOp represent different ways to assign a value

Constructors

NormalAssign

MultiplyAssign

DivideAssign

ModuloAssign

PlusAssign

MinusAssign

ShiftLeftAssign

SignedShiftRightAssign

USignedShiftRightAssign

AndAssign

XOrAssign

OrAssign

data BinOp

BinOp are all primitive binary operations

Constructors

And

Or

XOr

Eq

Less

LessEq

Greater

GreaterEq

InstanceOf

Multiply

Divide

Add

Subtract

Modulo

data UnOp

UnOp are all primitive unary operations

Constructors

Not

Neg

BitCompl
Tilde-Operator performs a bitwise complement

data IncrOrDecr

IncrOrDecr represent different ways of increment/decrement a field in an effectful way

Constructors

PreIncr

PostIncr

PreDecr

PostDecr