Compiler.ParserUtils

Description

This module contains definitions and some useful helpers for Compiler.Parser

Documentation

type Parser = Parsec Void String

makeSingleton :: Parser a -> Parser [a]

makeSingleton puts the result of a parser into an array

spaceConsumer :: Parser ()

spaceConsumer consumes spaces and comments

lexeme :: Parser a -> Parser a

lexeme is a wrapper for lexemes that consumes all spaces and comments after a lexeme

symbol :: String -> Parser String

Symbol

parens :: Parser a -> Parser a

parens parses something between parenthesis.

braces :: Parser a -> Parser a

braces parses something between parenthesis.

integer :: Parser Integer

integer parses an integer.

semicolon :: Parser ()

semicolon parses a semicolon.

comma :: Parser ()

comma parses a comma.

kword :: String -> Parser ()

kword parses a reserved word (keyword).

binOp :: String -> (a -> a -> a) -> Parser a -> Parser a

binOp defines recursive binary operation with a different parser for the left side.